requirements.txt
Django==4.1
psycopg2==2.9.3
requests==2.28.1
google-api-python-client==2.55.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
pyTelegramBotAPI==4.6.1
python-telegram-bot==13.13
prettytable==3.3.0
